[keyserver] Introduce getInitialReduxStateResponder
Summary:
[ENG-4751](https://linear.app/comm/issue/ENG-4751)
Depends on D9111

This diff creates the definition of the new `getInitialReduxStateResponder`. It takes `URLInfo` as the input, because we can generate it on the client, and that makes it more consistent with the rest of the responders. The output is modified from initial redux state in website responder:
- removed all top-level fields which were always set to a default (except for `commServicesAccessToken` which I assume will be filled in here in the future)
- instead of `UserStore` return `UserInfos` (the `inconsistencyReports` field is always set to default, so it will be set on client)
- keyserverInfo contains only `sessionID` and `updatesCurrentAsOf` because:
  - `cookie` and `lastCommunicatedPlatformDetails` are persisted
  - `connection` is reset to default after reload
  - `urlPrefix` is set to default (and will be probably persisted later)

Test Plan: `yarn flow`

Reviewers: inka, kamil, atul

Subscribers: