[web] Remove intial redux state from website responders
Summary:
[ENG-4752](https://linear.app/comm/issue/ENG-4752)
Depends on D9141

After adding the call to initial redux state we can finally remove it from `website-responders`. We need to add a new default web state that will make sure the app runs while we make the call to the keyserver for the initial state.

Test Plan: Load the web app when logged in and when logged out. Check that the app works and there are no errors in the console.

Reviewers: kamil, inka, atul

Subscribers: