[keyserver] Read calendar query directly from viewer.
Summary:
We were providing calendar query to a fetch from a spec as a separate parameter. Each time a value for this parameter was computed, it was taken from the viewer object. It means that we can simplify the code by reading the query inside specs that need it.

https://linear.app/comm/issue/ENG-4822/consider-using-calendarquery-from-viewer-inside-state-sync-specs

Depends on D9081

Test Plan: Modify server state and check if state check fixes client state.

Reviewers: kamil, bartek, inka

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9086