[web/keyserver] Handle different thread ids in url
Summary:
[ENG-4752](https://linear.app/comm/issue/ENG-4752)
Depends on D9147
Depends on D9146

We want to handle older links too. Previously the server just ignored errors if the link couldn't be converted (so e.g. it contained older ids). Now we convert the ids in the urlinfo to the new schema on the client and server always tries to convert them.

Test Plan:
Navigated to:
- `http://localhost:3000/comm/chat/thread/256%7C91890/` and `http://localhost:3000/comm/chat/thread/91890/` -> web app opened correct thread
- `http://localhost:3000/comm/chat/thread/pending/sidebar/256%7C91894/` and `http://localhost:3000/comm/chat/thread/pending/sidebar/91894/` -> the keyserver doesn't support navigating to pending threads right now, so the web app opened on the most recent thread, but there were no error
- `http://localhost:3000/comm/chat/thread/new/83884+86227/` -> web app opened on the chat creation menu with the selected users
- `http://localhost:3000/comm/settings/account/` -> web app opened on settings

Also run `yarn jest`

Reviewers: kamil, inka, atul

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9148