[lib] Create selectors for fetching data needed to call servers
Summary:
issue: https://linear.app/comm/issue/ENG-4683/refactor-useservercall-step-1-refactor-fetching-data-from-state
We currently fetch some data from state using serverCallStateSelector, and then bind this data to callServerEndpoint. But with multiple keyservers we won't know which keyserver is going to be called. Instead we will
fetch a collection of shape { [keyserverID]: {cookie, urlPrefix, ...} } and eventually pass it to the returned function, which will extract the data it needs. This is not very memory intensive, since we can create a
selector with createSelector, which memoizes the result and returns the reference to the same object as long as the result doesn't change (even in different components).

Test Plan:
Ran the selector, chcked that it returns the expected value, as defined by `ServersCallState`.
ran `yarn flow-all`

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9092