[web] Fix nested a tags
Summary:
issue: https://linear.app/comm/issue/ENG-2596/anchor-tag-contains-another-anchor-tag-on-web
Nested a tags cause errors. The only place where SidebarItem is used on web, is in ChatThreadListSidebar. The only purpose the two <a> tags serve is having an onClick method, that is the same in both cases -
useOnClickThread(threadInfo). So I propose to remove the outer tag, and instead make the underlying SidebarItem take as much space as possible

Test Plan:
Checked that pressing the sideabr item in the chat thread list navigates the user to thet sidebar.

Reviewers:
rohan, ginsu, michal, atul
Subscribers: