[web] Hide databaseModule behind a function
Summary:
Part of [ENG-4844](https://linear.app/comm/issue/ENG-4844)
When database will be used by all users (including anonymous ones), we need to make sure that it's never in an unitialized state. We hide it behind a getter function so we add the initialization code there in the later diffs.

Test Plan: `yarn flow`

Reviewers: kamil!, tomek, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9105