[native] introduce onClosed callback prop to BottomSheet component
Summary:
In subsequent diffs we will introduce some logic where we want to navigate back from the modal bottom sheet screen back to the screen we originally were before we opened/navigated to the modal bottom sheet screen when the bottomsheet closes. Unfortunately, the bottom sheet library does not have a callback prop for this. However, it does have a callback prop called `onChange` that gets triggered every time the sheet position changes. We can use this and set a condition where if the sheet position index is -1 (the sheet is closed) then we call the `onClosed` callback prop

https://gorhom.github.io/react-native-bottom-sheet/props/#onchange

Test Plan:
Passed in a callback which logged a message to the console as the `onClosed` prop and confirmed that the message appeared when the bottomsheet closed

{F746496}

Reviewers: atul, inka

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9098