[native] Memoize `swipeableThread` in `ChatThreadListItem`
Summary:
This prevents re-renders of `SwipeableThread`s. As of this diff we've reduced the time of `ChatThreadList` re-renders from ~600ms to ~70ms.

`ChatThreadListItem ` before D9102:
{F746742}

`ChatThreadListItem` before this diff:
{F746741}

`ChatThreadListItem` after this diff:
{F746740}

(This is the `chatThreadListItem` for Daily Updates)

---

Depends on D9102

Test Plan: Things continue to look/work as expected.

Reviewers: ginsu, tomek, rohan

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9103