[services] Delete base-image and sandbox mode
Summary:
No services use the C++ base-image anymore so it can be deleted. Also, removed leftovers of sandbox mode, which we decided to remove some time ago.

Depends on D9116, D9117

Test Plan: None, this is just a cleanup. Did a git grep though to make sure nothing is using it.

Reviewers: michal, varun, kamil, jon

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9137