[keyserver] Remove unused params in validation functions
Summary:
Depends on D9145

After the previous diff we can remove these "special case" parameters from validation functions. They were used only in `website-responders`. They were added because in `website-responders` the viewer can have old `platformDetails` while we now that after this request they will have the newest version (because we are serving it). So in this case we wanted to force the id conversion. But now the platform details are correctly updated during the call to `redux-state-responders`.

Test Plan: `yarn flow-all`

Reviewers: kamil, inka, atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9146