[native] Lift `searchCancelButton*` `Value`/`Node`s  to `ConnectedChatThreadList`
Summary:
(I initially had a huge diff that converted `ChatThreadList` to a functional component and added various memoizations. However, I realized that diff would be pretty difficult to review so I'm breaking down the refactor into hopefully easier to review pieces.)

This diff is just one step in the process of converting `ChatThreadList` into a functional component. When this work is done, we will be able to avoid a lot of re-rendering and hopefully improve performance quite a bit.

In subsequent diffs will
1. Lift up callbacks related to search
2. Lift up "search-related" JSX
3. etc.. continue turning `ChatThreadList` into a functional component so we can more precisely memoize things, particularly `ChatThreadListItem`s.

---

Depends on D9132

Test Plan:
Search experience continues to work as expected. Can search for users/threads/etc.

NOTE: Noticed an existing issue during testing that WAS NOT caused by this refactor: https://linear.app/comm/issue/ENG-4922/search-chats-x-icon-regression

Reviewers: ginsu, tomek, rohan

Reviewed By: ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9133