[native] Replace `scrollPos` instance variable w/ ref
Summary:
As part of converting `ChatThreadList` to a functional component we'll need to replace the `scrollPos` instance variable with a ref.

---

Depends on D9135

Test Plan:
Things seem to continue to work as expected. Haven't used `useRef` a whole lot, so would be good to get feedback from reviewers w/ more familiarity.

Reviewers:
ashoat, ginsu, rohan, tomek