[lib/native] lift useIsCurrentUserStaff hook into lib
Summary: To better test the light mode/color famlies on web, I wanted to create an interface to make it easy to toggle between light mode and dark mode. Since this will be longer work in progress process, I wanted to gate this interface so only Comm staff can see it. As a first step to this, I want to lift `useIsCurrentUserStaff` into `lib` so `web` can use this hook too

Test Plan: `flow` and confirmed that `useIsCurrentUserStaff` still returns the same value as it did when it lived in `native`

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9193