[native] Lift `search[Status/Text]` state to `ConnectedChatThreadList`
Summary:
(I initially had a huge diff that converted `ChatThreadList` to a functional component and added various memoizations. However, I realized that diff would be pretty difficult to review so I'm breaking down the refactor into hopefully easier to review pieces.)

This diff is just one step in the process of converting `ChatThreadList` into a functional component. When this work is done, we will be able to avoid a lot of re-rendering and hopefully improve performance quite a bit.

Test Plan: Search experience continued to work as expected.

Reviewers: ginsu, rohan, tomek

Reviewed By: ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9130