[keyserver] Print out a QR code to the terminal
Summary:
This diff handles printing out a QR code to the terminal - the idea is that this will ultimately be displayed on the keyserver. For now I've used temporary keys, the next two diffs will encode an
ephemeral AES-256 key and the ed25519 key.

Addresses [[ https://linear.app/comm/issue/ENG-4776/print-out-a-qr-code-to-the-terminal | ENG-4776 ]] and [[
https://linear.app/comm/issue/ENG-4773/set-a-condition-to-determine-whether-a-qr-code-should-be-shown-in-the | ENG-4773 ]].

Depends on D9089

Test Plan: Confirmed that toggling the flag ensures that the QR code prints out to the terminal

Reviewers: atul, ginsu, varun, patryk

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9180