[keyserver] Generate an AES256 key (ephemeral) to encode
Summary:
The first key we want to encode into the QR code is the AES-256 ephemeral key. The next diff will be the device id.

Depends on D9180

Addresses [[ https://linear.app/comm/issue/ENG-4778/generate-an-aes256-key-ephemeral-for-the-new-device | ENG-4778 ]]

Test Plan:
Confirmed that the ephemeral key is encoded in the URL. Essentially repeating the test plan from D9093

{F757317}

Reviewers: atul, ginsu, varun, patryk

Reviewed By: varun

Subscribers: anunay, ashoat, tomek

Differential Revision: https://phab.comm.dev/D9181