[native] Lift `floatingAction` to `ConnectedChatThreadList`
Summary:
This diff is just one step in the process of converting ChatThreadList into a functional component. When this work is done, we will be able to avoid a lot of re-rendering and hopefully improve performance quite a bit.

NOTE: Also snuck in some cleanup to make diff where we merge into single component a bit more readable.
---

Depends on D9174

Test Plan: ChatThreadList + search experience continue to work as expected.

Reviewers: ginsu, tomek, rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9183