[native] Lift `onTabPress` to `ConnectedChatThreadList`
Summary:
This diff is just one step in the process of converting ChatThreadList into a functional component. When this work is done, we will be able to avoid a lot of re-rendering and hopefully improve performance quite a bit.

---

Depends on D9184

Test Plan:
ChatThreadList + search experience continue to work as expected.

NOTE: Specifically tried tapping the tab and ensuring the scrollToTop functionality worked as expected.

Reviewers: ginsu, tomek, rohan

Reviewed By: ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9188