[native] use stringForUserExplicit function to determine username text for profiles
Summary:
At this point the `UserProfile` component needed a `userInfo` of type `AccountUserInfo` to render. However, we need to consider the case when we view a profile and we don't know the username on the client (an anonymous user). `stringForUserExplicit` is a function that handles this case for us and this also allows us to switch the `userInfo` prop type to be the more permissive `UserInfo` type

Depends on D9094

Test Plan:
Please see screenshot below, and confirmed that UserInfos with known usernames still show the correct username

{F751982}

Reviewers: atul, inka

Reviewed By: inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9119