[lib] introduce threadInfosSelectorForThreadType selector
Summary: The next step for user profiles is to select the threadInfo of the thread between the viewer and the user who's profile is being viewed. For user profiles there are two types of thread types we need to consider: `PERSONAL` and `PRIVATE`.

Test Plan: Ran the following code in the `UserProfileBottomSheet` component and confirmed I was returning the correct information

Reviewers:

Subscribers: