[native] Remove `renderSearch` in `ChatThreadList` and inline `ChatThreadListSearch` directly
Summary:
Instead of passing a function to create `ChatThreadListSearch` we can inline the component directly.

As part of this we also replace `additionalProps` w/ more precise `innerSearchAutoFocus` and `innerSearchActive` props.

---

Depends on D9207

Test Plan:
Search experience continues to look and work as expected. Set logs in all expected callbacks to ensure that they fired as expected. Made sure that x-icon and "Cancel" button continue to look and behave as before.

Reviewers:
ginsu, tomek, rohan