[native] Forward `searchInputRef` from `ChatThreadList` to `ChatThreadListSearch`
Summary:
Forward ref from `ChatThreadList` to `ChatThreadListSearch`... which will in turn forward it to the inner `Search` component.

---

Depends on D9208

Test Plan:
Search experience continues to look and work as expected. Set logs in all expected callbacks to ensure that they fired as expected. Made sure that x-icon and "Cancel" button continue to look and behave as before.

Reviewers:
tomek, ginsu, rohan