[native] Memoize `flatList` in `ChatThreadList`
Summary:
We want to ensure that the `FlatList` renders only when absolutely necessary. This helps "isolate" the `FlatList` from state that may trigger re-renders of eg `fixedSearch`. For example, we want to isolate `flatList` from changes to `searchText` which would always trigger re-render of `fixedSearch` which would cause the entirety of `chatThreadList` to re-render. (In practicce a change to `searchText would probably change eg `threadSearchResults` which would in turn change `partialData` which would trigger flatList to expectedly re-render.. but just for demonstration).

---

Depends on D9210

Test Plan: Search experience continues to look and work as expected. Set logs in all expected callbacks to ensure that they fired as expected. Made sure that x-icon and "Cancel" button continue to look and behave as before.

Reviewers: ginsu, tomek, rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9211