[native] Consume `useThreadListSearch` in `ChatThreadList`
Summary:
Huge shoutout @Jacek for giving us the gift of `useThreadListSearch` which we're able to just seamlessly drop into `ChatThreadList`. This further helps reduce the number of `ChatThreadList` re-renders.

We're getting closer, but there's still some more work to get to no re-renders of `ChatThreadList` when we navigate to MessageList from ThreadList...

---

Depends on D9211

Test Plan: Search experience continues to look and work as expected. Set logs in all expected callbacks to ensure that they fired as expected. Made sure that x-icon and "Cancel" button continue to look and behave as before.

Reviewers: ginsu, tomek, rohan

Reviewed By: rohan

Subscribers: ashoat, jacek

Differential Revision: https://phab.comm.dev/D9212