[native] Memoize `ThreadAncestorsLabel`
Summary:
Looking at the profiler flamegraph for the "navigate to thread w/ varun" flow, there are a lot of `ThreadAncestorsLabel` "slivers" that seem like they might be causing other things to re-render? Unfortunately the profiler isn't super clear so I'm kind of memoizing this on a whim tbh.

---

Depends on D9212

Test Plan:
`ThreadAncestorsLabel` continues to look as expected.

Reviewers:
ginsu, tomek, rohan