[native] Remove `boundChatListData` from `useThreadListSearch` dep list
Summary:
There's no reason for `boundChatListData` to be an argument to `useThreadListSearch` as it's only included in a dep list where it's also not needed.

When `boundChatListData` updated it forced an update of `threadSearchResults` and `usersSearchResults` which led to `listData` getting recomputed THREE times. After this change it only gets recomputed once and causes `ChatThreadList` to re-render once.

Before:

After:

---

Depends on D9214

Test Plan: Things continue to work as expected and the "navigate to thread with varun" flow is now a lot faster.

Reviewers: ginsu, tomek, rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D9215