[terraform] Deploy Tunnelbroker DynamoDB table
Summary: It turns out that this table was missing on AWS. Interestingly, @kamil noticed that tunnelbroker actually hasn't been saving anything to dynamodb and the test is broken. This is why I didn't notice that the table was missing - tunnelbroker wasn't failing without the table.

Test Plan: Terraform plan + terraform apply on staging created the table.

Reviewers: kamil, jon

Reviewed By: kamil

Subscribers: ashoat, tomek, kamil

Differential Revision: https://phab.comm.dev/D9216