[web] Migrate cookies to redux
Summary:
[ENG-4767](https://linear.app/comm/issue/ENG-4746)

We want to migrate the browser cookies to redux. An explanation why we decided is in [ENG-4347](https://linear.app/comm/issue/ENG-4347) and on [Notion](https://www.notion.so/commapp/Browser-cookies-CORS-84fab494c7e542569d1b5a1a15e22807). We do the migration on the client after setting the cookie to `httpOnly: false` on the keyserver. This is normally not recommended because of XSS, but if our end goal is moving the cookie to redux, it will be accesible to js anyway, so there shouldn't be any issue with this. If the attacked can execute any js on the client they can still make authenticated calls to the keyserver/services (for services we keep `commAccessToken` in redux anyway).

Test Plan: Run the migration, check if the cookie was stored in redux

Reviewers: inka, kamil, atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9287