[keyserver] Stop sending http cookies
Summary:
[ENG-4747](https://linear.app/comm/issue/ENG-4747)
Depends on D9288

We want to stop sending browser cookies on requests from the keyserver. The only exception is for `website-responders` where we send back any cookies we get. This is done so that they are updated to `httpOnly: false` and so older clients that will update to redux cookies can migrate.

Test Plan:
- Check that the migration still works
- Check that the web app still works
- Check that the cookie lifespan is still updated

Reviewers: inka, kamil, atul

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D9290