[native] commCoreModule methods for getting, setting, and clearing the commServicesAccessToken
Summary: we need the access token to be available in the SecureStore on native for NSE.

Test Plan: from account-selector.js i called the setter and getter, getting back the token i set in the secure store. then i cleared the token and tried to get it again, getting an error as expected.

Reviewers: marcin!

Subscribers: